Remove all the code

+++ src/string_buffer.h
@@ -6,13 +6,13 @@
 #include <stdio.h>
 
 typedef struct {
-    char *data;     // pointer to buffer
-    int len;        // current length
-    int cap;        // allocated capacity
+    char *data;
+    int len;
+    int cap;
 } StringBuffer;
 
 void sb_init(StringBuffer *sb, int initial_cap) {
-    sb->data = malloc(initial_cap);
+    sb->data = malloc((unsigned long)initial_cap);
     sb->len = 0;
     sb->cap = initial_cap;
 }
@@ -25,13 +25,13 @@ void sb_free(StringBuffer *sb) {
 }
 
 void sb_append_string(StringBuffer *sb, const char *str) {
-    int str_len = strlen(str);
+    int str_len = (int)strlen(str);
     if (sb->len + str_len >= sb->cap) {
         int new_cap = sb->cap * 2;
         while (new_cap <= sb->len + str_len) {
             new_cap *= 2;
         }
-        char *tmp = realloc(sb->data, new_cap);
+        char *tmp = realloc(sb->data, (unsigned long)new_cap);
         if (!tmp) {
             fprintf(stderr, "Out of memory\n");
             exit(1);
@@ -40,14 +40,14 @@ void sb_append_string(StringBuffer *sb, const char *str) {
         sb->cap = new_cap;
     }
 
-    memcpy(sb->data + sb->len, str, str_len);
+    memcpy(sb->data + sb->len, str, (unsigned long)str_len);
     sb->len += str_len;
 }
 
 void sb_append_char(StringBuffer *sb, char c) {
     if (sb->len + 1 >= sb->cap) {
         sb->cap *= 2;
-        sb->data = realloc(sb->data, sb->cap);
+        sb->data = realloc(sb->data, (unsigned long)sb->cap);
     }
     sb->data[sb->len++] = c;
 }
